requirements.txt
python>=3.8
requests==2.28.2
pandas==2.0.3
numpy==1.24.3
geopy==2.3.0
googlemaps==4.10.0
pycountry==22.3.5
matplotlib==3.7.1
seaborn==0.12.2
folium==0.14.0
networkx==3.1
pyvis==0.3.2
deep-translator==1.11.4
opencc-python-reimplemented==0.1.7
iso3166==2.1.1
statsmodels==0.14.0
google-cloud-translate==2.11.1
langdetect==1.0.9
pyyaml==6.0.1
google-api-python-client==2.104.0
